Changed project structure
Moved internal functions to `src/iapws97/regions.rs`.
Public functions to `src/iapws97.rs`
