+++ Cargo.toml
@@ -1,12 +1,14 @@
 [package]
 name = "rust-steam"
-version = "0.1.9"
+version = "0.2.0"
 edition = "2021"
 description = "A Rust implementation of the IAPWS-IF97 steam tables"
 repository = "https://github.com/marciorvneto/rusteam"
 license = "MIT"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 float-cmp = "0.9.0"
+
+[lib]
+name = "rust_steam"
+path = "src/IAPWS97.rs"
